refactor: move experiment/designspec API helpers up to experiment-utils.ts
