+++ src/app/experiments/create/experiment-form/experiment-form-types.ts
@@ -1,26 +1,17 @@
 import {
-  AnyFrequentistDesignSpecExperimentType,
   Arm,
-  CMABExperimentSpec,
-  CMABExperimentSpecExperimentType,
   ContextType,
   CreateExperimentResponse,
-  DesignSpec,
-  ExperimentConfig,
   FieldMetadata,
   Filter,
-  GetExperimentResponse,
   GetFiltersResponseElement,
   GetMetricsResponseElement,
   MABExperimentSpec,
-  MABExperimentSpecExperimentType,
-  OnlineFrequentistExperimentSpec,
-  OnlineFrequentistExperimentSpecExperimentType,
   PowerResponse,
-  PreassignedFrequentistExperimentSpec,
   PreassignedFrequentistExperimentSpecExperimentType,
 } from '@/api/methods.schemas';
 import { ErrorType } from '@/services/orval-fetch';
+import type { ExperimentType } from '@/services/experiment-utils';
 
 export type ContextVariableType = ContextType;
 
@@ -80,8 +71,6 @@ export type MetricWithMDE = {
   metric: GetMetricsResponseElement;
   mde: string; // desired minimum detectable effect as a percentage of the metric's baseline value
 };
-
-export type ExperimentType = DesignSpec['experiment_type'];
 
 // Defines the entirety of the editable data collected via this wizard flow.
 export type ExperimentFormData = {
@@ -153,47 +142,5 @@ export type ExperimentScreenId =
   | 'summarize-freq'
   | 'summarize-bandit';
 
-// Define the type alias using imported types
-export const isFreqExperimentType = (
-  experimentType?: ExperimentType,
-): experimentType is AnyFrequentistDesignSpecExperimentType =>
-  experimentType === PreassignedFrequentistExperimentSpecExperimentType.freq_preassigned ||
-  experimentType === OnlineFrequentistExperimentSpecExperimentType.freq_online;
-
-export const isClusteredExperiment = (data: ExperimentFormData): boolean =>
+export const isClusteredExperimentFormData = (data: ExperimentFormData): boolean =>
   data.experimentType === PreassignedFrequentistExperimentSpecExperimentType.freq_preassigned && !!data.clusterKey;
-
-export const isCmabExperimentType = (
-  experimentType?: ExperimentType,
-): experimentType is CMABExperimentSpecExperimentType =>
-  experimentType === CMABExperimentSpecExperimentType.cmab_online;
-
-export const isBanditExperimentType = (experimentType?: ExperimentType): experimentType is BanditExperimentType =>
-  experimentType === MABExperimentSpecExperimentType.mab_online ||
-  experimentType === CMABExperimentSpecExperimentType.cmab_online;
-
-export const isFrequentistSpec = (
-  spec: DesignSpec | undefined,
-): spec is OnlineFrequentistExperimentSpec | PreassignedFrequentistExperimentSpec =>
-  !!spec && isFreqExperimentType(spec.experiment_type);
-
-export const isFreqPreassignedSpec = (spec: DesignSpec | undefined): spec is PreassignedFrequentistExperimentSpec =>
-  !!spec && spec.experiment_type === PreassignedFrequentistExperimentSpecExperimentType.freq_preassigned;
-
-export const isClusteredPreassignedSpec = (
-  spec: DesignSpec | undefined,
-): spec is PreassignedFrequentistExperimentSpec => isFreqPreassignedSpec(spec) && !!spec.cluster_key;
-
-export function isMabSpec(spec: DesignSpec | undefined): spec is MABExperimentSpec {
-  return !!spec && spec.experiment_type === MABExperimentSpecExperimentType.mab_online;
-}
-
-export function isCmabSpec(spec: DesignSpec | undefined): spec is CMABExperimentSpec {
-  return !!spec && spec.experiment_type === CMABExperimentSpecExperimentType.cmab_online;
-}
-
-export const isBanditSpec = (spec: DesignSpec | undefined): spec is MABExperimentSpec | CMABExperimentSpec =>
-  isMabSpec(spec) || isCmabSpec(spec);
-
-export const isCmabExperiment = (experiment: GetExperimentResponse | ExperimentConfig | undefined): boolean =>
-  !!experiment && isCmabSpec(experiment.design_spec);
