+++ src/services/experiment-utils.ts
@@ -1,4 +1,20 @@
 import type { ExperimentStatus } from '@/components/features/experiments/types';
+import {
+  AnyFrequentistDesignSpecExperimentType,
+  CMABExperimentSpec,
+  CMABExperimentSpecExperimentType,
+  DesignSpec,
+  ExperimentConfig,
+  GetExperimentResponse,
+  MABExperimentSpec,
+  MABExperimentSpecExperimentType,
+  OnlineFrequentistExperimentSpec,
+  OnlineFrequentistExperimentSpecExperimentType,
+  PreassignedFrequentistExperimentSpec,
+  PreassignedFrequentistExperimentSpecExperimentType,
+} from '@/api/methods.schemas';
+
+export type ExperimentType = DesignSpec['experiment_type'];
 
 export const getExperimentStatus = (startDate: string, endDate: string): ExperimentStatus => {
   const now = new Date();
@@ -13,3 +29,46 @@ export const getExperimentStatus = (startDate: string, endDate: string): Experim
     return 'current';
   }
 };
+
+export const isFreqExperimentType = (
+  experimentType?: ExperimentType,
+): experimentType is AnyFrequentistDesignSpecExperimentType =>
+  experimentType === PreassignedFrequentistExperimentSpecExperimentType.freq_preassigned ||
+  experimentType === OnlineFrequentistExperimentSpecExperimentType.freq_online;
+
+export const isCmabExperimentType = (
+  experimentType?: ExperimentType,
+): experimentType is CMABExperimentSpecExperimentType =>
+  experimentType === CMABExperimentSpecExperimentType.cmab_online;
+
+export const isBanditExperimentType = (
+  experimentType?: ExperimentType,
+): experimentType is MABExperimentSpecExperimentType | CMABExperimentSpecExperimentType =>
+  experimentType === MABExperimentSpecExperimentType.mab_online ||
+  experimentType === CMABExperimentSpecExperimentType.cmab_online;
+
+export const isFrequentistSpec = (
+  spec: DesignSpec | undefined,
+): spec is OnlineFrequentistExperimentSpec | PreassignedFrequentistExperimentSpec =>
+  !!spec && isFreqExperimentType(spec.experiment_type);
+
+export const isFreqPreassignedSpec = (spec: DesignSpec | undefined): spec is PreassignedFrequentistExperimentSpec =>
+  !!spec && spec.experiment_type === PreassignedFrequentistExperimentSpecExperimentType.freq_preassigned;
+
+export const isClusteredPreassignedSpec = (
+  spec: DesignSpec | undefined,
+): spec is PreassignedFrequentistExperimentSpec => isFreqPreassignedSpec(spec) && !!spec.cluster_key;
+
+export function isMabSpec(spec: DesignSpec | undefined): spec is MABExperimentSpec {
+  return !!spec && spec.experiment_type === MABExperimentSpecExperimentType.mab_online;
+}
+
+export function isCmabSpec(spec: DesignSpec | undefined): spec is CMABExperimentSpec {
+  return !!spec && spec.experiment_type === CMABExperimentSpecExperimentType.cmab_online;
+}
+
+export const isBanditSpec = (spec: DesignSpec | undefined): spec is MABExperimentSpec | CMABExperimentSpec =>
+  isMabSpec(spec) || isCmabSpec(spec);
+
+export const isCmabExperiment = (experiment: GetExperimentResponse | ExperimentConfig | undefined): boolean =>
+  !!experiment && isCmabSpec(experiment.design_spec);
